% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_from_quarter.R
\name{month_from_quarter}
\alias{month_from_quarter}
\title{Get the last month in a numeric calendar quarter}
\usage{
month_from_quarter(quarter)
}
\arguments{
\item{quarter}{numeric between 1-4 indicating the quarter of a calendar year}
}
\description{
Takes a numeric calendar quarter value to produce the last month in that quarter
}
