% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_source.R
\name{prior_source}
\alias{prior_source}
\title{Find a specified recent weblink to download zip files from}
\usage{
prior_source(origin, year_quarter)
}
\arguments{
\item{origin}{whether CSV comes from mortgage and landlord "mlp" or civil statistics "civil".}

\item{year_quarter}{string of year and quarter of interest, in the format "YYYY-Q"}
}
\description{
Starts from root directory to find prior civils pages
}
