% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_previous.R
\name{read_previous}
\alias{read_previous}
\title{Read in CSV of interest as a data.table}
\usage{
read_previous(csv_name, data_type, year_quarter)
}
\arguments{
\item{csv_name}{csv name as a string, choose from "court", "la", "map", and "monthly" for mortgage and landlord possession, and "workload", "jr", "timeliness" and "progression" for civil}

\item{data_type}{whether CSV comes from mortgage and landlord "mortgage" or civil statistics "civil".}

\item{year_quarter}{string of year and quarter of interest, in the format "YYYY-Q"}
}
\description{
Allows reading in of the published Civil Justice and Mortgage and Landlord Possession data CSVs for a specific date in an automated way
}
