% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_source.R
\name{web_source}
\alias{web_source}
\title{Find most recent weblink to download zip files from}
\usage{
web_source(origin)
}
\arguments{
\item{origin}{whether CSV comes from mortgage and landlord "mlp" or civil statistics "civil".}
}
\description{
Starts from root directory to find most recent civils pages
}
